+++ komis-backend/CarShop/src/main/java/com/TAB/CarShop/Entities/Showroom.java
@@ -16,7 +16,7 @@ public class Showroom {
     @Id
     @Column(name = "showroom_id", nullable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int showroom_id;
+    private long showroom_id;
 
     @Column(name = "address", nullable = false)
     private String address;
